Move answer submit js to webpack
